pyproject.toml
[project]
name = "nviro-airflow-parsing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=7.0.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "azure-storage-blob>=12.26.0",
    "boto3>=1.40.16",
    "faker>=37.5.3",
    "google-cloud-storage>=3.3.0",
    "great-expectations>=0.18.8",
    "httpx>=0.28.1",
    "markdownify>=1.2.0",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "paramiko<3.0.0",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pydantic>=2.11.7",
    "pymongo>=4.14.1",
    "pymupdf>=1.26.4",
    "pymupdf4llm>=0.0.27",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sentry-sdk>=2.35.0",
    "sqlalchemy>=2.0.43",
    "sshtunnel>=0.4.0",
    "psutil>=5.9.0",
    "tqdm>=4.66.0",
    "pdfplumber>=0.11.7",
    "pypdf>=6.1.1",
    "langchain-core>=0.3.79",
    "neo4j>=6.0.2",
    "langchain-openai>=0.3.35",
    "langchain-community>=0.3.31",
]
